requirements.txt
crewai
crewai_tools
langchain-google-genai
langchain-groq
py-dotenv
langchain